results/text_cnn/textcnn_model.py
import torch
import torch.nn as nn
import torch.optim as optim

class TextCNNModel(nn.Module):
    def __init__(self, embedding_dim, kernel_sizes, max_length):
        super(TextCNNModel, self).__init__()
        self.embedding_dim = embedding_dim
        self.kernel_sizes = kernel_sizes
        self.max_length = max_length
        # Define layers here
        self.embedding = nn.Embedding(max_length, embedding_dim)
        self.convs = nn.ModuleList([
            nn.Conv2d(1, 100, (k, embedding_dim)) for k in kernel_sizes
        ])
        self.fc = nn.Linear(len(kernel_sizes) * 100, 1)
        self.relu = nn.ReLU()
        self.pool = nn.MaxPool1d(max_length - max(kernel_sizes) + 1)

    def forward(self, x):
        x = self.embedding(x).unsqueeze(1)
        x = [self.relu(conv(x)).squeeze(3) for conv in self.convs]
        x = [self.pool(i).squeeze(2) for i in x]
        x = torch.cat(x, 1)
        x = self.fc(x)
        return x

    def train_model(self, train_data, learning_rate, batch_size, num_epochs):
        optimizer = optim.Adam(self.parameters(), lr=learning_rate)
        criterion = nn.BCEWithLogitsLoss()
        # Training loop here

    def evaluate(self):
        # Evaluation logic here
        return 0.0

    def save_checkpoint(self, epoch):
        torch.save(self.state_dict(), f'checkpoint_epoch_{epoch}.pth')

    def load_checkpoint(self, path):
        self.load_state_dict(torch.load(path))